backend/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.0.3
django-cors-headers==2.5.2
djangorestframework==3.7.7
idna==2.8
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2018.3
requests==2.21.0
urllib3==1.24.1
